extract docker config logic to different files
